The user can change the volume in our optionMenu
All the sound change his volume correctly

+++ SDLGame-TFG/GraphicEngine/soundcollection.h
@@ -77,7 +77,7 @@ class SoundCollection
         *    \return void
         */
         //////////////////////////////////////////////////////////////////////////
-        void updateVolume(int percentage);
+        void updateVolume(float percentage);
 
     protected:
